Refactored js

+++ trunk/webapp/public/javascripts/shoppinglist.mod.households.js
@@ -35,27 +35,40 @@ var Shoppinglist = Shoppinglist || {};
         return Shoppinglist.helper.talk_to_controller('fetchinvitations', 'invitations', callback);
     };
 
-    update_existing_view = function (data) {
+    /**
+     * Creates html for existing households
+     * @method update_existing_view
+     * @param {household} data
+     * @return {boolean} true
+     */
+    update_existing_view = function (data, callback) {
         var tmpHtml = new Shoppinglist.helper.HtmlString(),
             i = 0,
             len = 0,
             current_item = null;
-        
-        tmpHtml.add('<ul>');
+
+        if (data.households.length === 0) {
+            return false;
+        }
+
+        tmpHtml.add('<h2 class="base">My Households</h2>');
+        tmpHtml.add('<ul class="menu">');
         for (i = 0, len = data.households.length; i < len; i++) {
             current_item = data.households[i];
             tmpHtml.add('<li>');
-            tmpHtml.add('<img src="images/icon_house.png" height="20" width="20" border="0">');
+            tmpHtml.add('<img src="images/icon_house.png" height="20" width="20" class="base">');
             tmpHtml.add('<a href="controller_proxy.php?controller=fetchshoppinglists&amp;hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">' + current_item.name + '</a>');
 
             if (parseInt(current_item.isOwner, 10) === 1) {
                 tmpHtml.add('<span class="actions">');
-                tmpHtml.add('<a class="delete" title="delete" onclick="return confirm(\'Are you sure?\')" href="controller_proxy.php?controller=deletehousehold&amp;hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">');
-                tmpHtml.add('<img src="images/icon_delete.png" height="20" width="20" border="0"');
-                tmpHtml.add('</a>&nbsp;&nbsp;&nbsp;');
                 tmpHtml.add('<a class="invite" title="invite" href="?hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">');
-                tmpHtml.add('<img src="images/icon_invite.png" height="20" width="20" border="0"');
+                tmpHtml.add('<img src="images/icon_invite.png" height="20" width="20" class="base" />');
                 tmpHtml.add('</a>');
+                
+                tmpHtml.add('<a class="delete" title="delete" onclick="return confirm(\'Are you sure?\')" href="controller_proxy.php?controller=deletehousehold&amp;hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">');
+                tmpHtml.add('<img src="images/icon_delete.png" height="20" width="20" class="base" />');
+                tmpHtml.add('</a>');
+
                 tmpHtml.add('</span>');
             }
             tmpHtml.add('</li>');
@@ -64,26 +77,53 @@ var Shoppinglist = Shoppinglist || {};
 
         $('.bdExisting', $ctx).html(tmpHtml.toString());
 
+        callback();
+        
+        return true;
     };
 
-    update_invitations_view = function (data) {
+    /**
+     * Creates html for pending household invitations
+     * @method update_invitations_view
+     * @param {invitations} data
+     * @return {boolean} true
+     */
+    update_invitations_view = function (data, callback) {
 
         var tmpHtml = new Shoppinglist.helper.HtmlString(),
             i = 0,
             len = 0,
             current_item = null;
 
-        tmpHtml.add('<ul>');
+        if (data.invitations.length === 0) {
+            return false;
+        }
+
+        tmpHtml.add('<h2 class="base">Pending Invitations</h2>');
+        tmpHtml.add('<ul class="menu">');
         for (i = 0, len = data.invitations.length; i < len; i++) {
             current_item = data.invitations[i];
             tmpHtml.add('<li>');
-            tmpHtml.add('<img src="images/icon_house.png" height="20" width="20" border="0">');
-            tmpHtml.add('<a href="controller_proxy.php?controller=fetchshoppinglists&amp;hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">' + current_item.householdName + '</a>');
+            tmpHtml.add('<img src="images/icon_house.png" height="20" width="20" class="base" />');
+            // use the link with household name to show details
+            tmpHtml.add('<a href="#" hid="' + current_item.householdId + '">' + current_item.householdName + '</a>');
+            tmpHtml.add('<span class="actions">');
+            tmpHtml.add('<a class="accept" title="accept" href="controller_proxy.php?controller=deletehousehold&amp;hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">');
+            tmpHtml.add('Y');
+            tmpHtml.add('</a>');
+            tmpHtml.add('<a class="decline" title="decline" onclick="return confirm(\'Are you sure?\')" href="?hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">');
+            tmpHtml.add('N');
+            tmpHtml.add('</a>');
+            tmpHtml.add('</span>');
+
             tmpHtml.add('</li>');
         }
         tmpHtml.add('</ul>');
 
         $('.bdInvitations', $ctx).html(tmpHtml.toString());
+
+        callback();
+        return true;
     };
 
     // TODO: prepare is common function used for households, shoppinglist and items. Apply OOP!!!
@@ -120,16 +160,21 @@ var Shoppinglist = Shoppinglist || {};
         prepare($ctx);
 
         fetch_invitations_by_user_email(function (data) {
-           update_invitations_view(data);
+            update_invitations_view(data, function () {
+                $('.bdInvitations', $ctx).removeClass('bdHidden');
+            });
         });
 
         $ctx.bind('dataChanged', function () {
 
             fetch_households_by_user_id(function (data) {
    
-                update_existing_view(data);
+                update_existing_view(data, function () {
+                    $('.bdExisting', $ctx).removeClass('bdHidden');
+                });
             });
         });
+
 
 
         $ctx.delegate('.actions a.delete', 'click', function () {
@@ -171,6 +216,16 @@ var Shoppinglist = Shoppinglist || {};
 
                 }
             });
+            return false;
+        });
+
+        $ctx.delegate('.actions a.accept', 'click', function () {
+            console.info('accept');
+            return false;
+        });
+
+        $ctx.delegate('.actions a.decline', 'click', function () {
+            console.info('decline');
             return false;
         });
